# docs/swagger.yaml
definitions:
  model.APIError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request
        type: string
      requestId:
        example: 7db692fa-c767-468f-af3d-9231b0f88c69
        type: string
      timestamp:
        example: "2025-10-05T13:45:00Z"
        type: string
    type: object
...
  model.ContestParticipant:
    properties:
      contestId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      inviteId:
        type: string
      joinedAt:
        type: string
      maxSquares:
        type: integer
      role:
        $ref: '#/definitions/model.ParticipantRole'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.ContestSwagger:
    properties:
      awayTeam:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      homeTeam:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        type: string
      quarterResults:
        items:
          $ref: '#/definitions/model.QuarterResult'
        type: array
      squares:
        items:
          $ref: '#/definitions/model.Square'
        type: array
      status:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      xLabels:
        items:
          type: integer
        type: array
      yLabels:
        items:
          type: integer
        type: array
    type: object
...
  model.ParticipantRole:
    enum:
    - owner
    - participant
    - viewer
    type: string
    x-enum-varnames:
    - ParticipantRoleOwner
    - ParticipantRoleParticipant
    - ParticipantRoleViewer
  model.QuarterResult:
    properties:
      awayTeamScore:
        type: integer
      contestId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      homeTeamScore:
        type: integer
      id:
        type: string
      quarter:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
      winner:
        type: string
      winnerCol:
        type: integer
      winnerName:
        type: string
      winnerRow:
        type: integer
    type: object
...
  model.Square:
    properties:
      col:
        type: integer
      contestId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      owner:
        type: string
      ownerName:
        type: string
      row:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
      value:
        type: string
    type: object
...
  model.UpdateParticipantRequest:
    properties:
      maxSquares:
        maximum: 100
        minimum: 0
        type: integer
      role:
        enum:
        - participant
        - viewer
        type: string
    type: object
...
info:
  contact: {}
  description: API for squares.maxstash.io
  title: Squares API
  version: 1.0.0
paths:
...
  /contests/{id}/participants:
    get:
      description: Returns all participants and their roles. Any participant can view.
        Public contests allow any authenticated user
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ContestParticipant'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Get all participants for a contest
      tags:
      - participants
  /contests/{id}/participants/{userId}:
    delete:
      description: Owner removes a participant, or a participant removes themselves;
        the removed participant's squares are cleared. The owner cannot be removed.
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Remove a participant from a contest
      tags:
      - participants
    patch:
      consumes:
      - application/json
      description: Owner updates a participant's role or max squares
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Update details
        in: body
        name: participant
        required: true
        schema:
          $ref: '#/definitions/model.UpdateParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ContestParticipant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Update a participant's role or square limit
      tags:
      - participants
...
  /contests/me:
    get:
      description: Returns all contests where the authenticated user is a participant
      parameters:
      - description: Filter contests by name (case-insensitive)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ContestSwagger'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Get all contests the user participates in
      tags:
      - participants
...
  /ws/contests/{id}:
    get:
      description: Establishes a persistent WebSocket connection to receive real-time
        updates for a specific contest
      parameters:
      - description: Contest ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "101":
          description: WebSocket connection upgraded
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIError'
      security:
      - BearerAuth: []
      summary: Connect to WebSocket for real-time contest updates
      tags:
      - ws
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"